chore: move func extractFromTiptap
